+++ packages/connect-wallet-modal/src/components/EIP6963WalletsList/ArrowBack.tsx
@@ -1,0 +1,37 @@
+import React from 'react';
+
+const ArrowBack = React.forwardRef(function ArrowBack(
+  props: React.SVGProps<SVGSVGElement>,
+  svgRef?: React.Ref<SVGSVGElement>,
+) {
+  return (
+    <svg
+      width="14"
+      height="11"
+      viewBox="0 0 14 11"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      ref={svgRef}
+      {...props}
+    >
+      <path
+        fill="none"
+        d="M13 5.5L1 5.5"
+        stroke="currentColor"
+        strokeWidth="2"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+      <path
+        fill="none"
+        d="M5.5 10L1 5.5L5.5 1"
+        stroke="currentColor"
+        strokeWidth="2"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+});
+
+export { ArrowBack };