+++ packages/core-react/src/index.ts
@@ -5,3 +5,5 @@ export * from './context/reefKnotContext';
 export * from './context/reefKnotProvider';
 export { ReefKnotModalContextProvider } from './context/reefKnotModalContext';
 export type { ReefKnotModalContextValue } from './context/reefKnotModalContext';
+export { providersStore } from './eip6963';
+export type { EIP6963ProviderDetail } from './eip6963';