+++ packages/ui-react/src/components/modal/ModalStyles.tsx
@@ -85,6 +85,7 @@ export const ModalHeaderStyle = styled.div<{
     min-height: 32px;
     display: flex;
     align-items: flex-start;
+    gap: ${spaceMap.md}px;
     margin-bottom: ${$short ? -spaceMap.md : 0}px;
     padding: ${spaceMap.xl}px ${spaceMap.xxl}px;
     font-size: ${fontSizesMap.md}px;
@@ -117,7 +118,7 @@ export const ModalTitleStyle = styled.div<{
     margin-right: ${$center && !$withCloseButton ? spaceMap.xxl : '0'}px;
     padding-top: ${$withTitleIcon ? spaceMap.sm : '0'}px;
     padding-left: 0;
-    padding-right: ${spaceMap.sm}px;
+    padding-right: 0;
     flex-grow: 1;
     align-self: center;
 
@@ -189,12 +190,11 @@ export const ModalBackStyle = styled(ButtonIcon).attrs({
   color: 'secondary',
   variant: 'ghost',
   size: 'xs',
+  square: true,
 })`
-  ${({ theme: { colors, spaceMap } }) => css`
-    color: ${colors.textSecondary};
-    flex-shrink: 0;
-    margin: 0 ${spaceMap.sm}px 0 -6px;
-    border-radius: 50%;
-    background: transparent !important;
-  `}
+  margin: 0 0 0 -6px;
+  color: ${({ theme }) => theme.colors.textSecondary};
+  flex-shrink: 0;
+  border-radius: 50%;
+  background: transparent !important;
 `;