+++ packages/ui-react/src/icons/index.ts
@@ -1,3 +1,4 @@
 export { default as ArrowBack } from './ArrowBack';
 export { default as Check } from './Check';
 export { default as Close } from './Close';
+export { default as Warning } from './Warning';