+++ packages/wallets/bitkeep/src/index.ts
@@ -17,6 +17,7 @@ export const Bitget: WalletAdapterType = ({ providersStore }) => ({
   walletName: name,
   walletId: id,
   type: injected.type,
+  rdns,
   icon: WalletIcon,
   detector: () => isProviderExistsEIP6963(providersStore, rdns),
   downloadURLs: {