+++ packages/wallets/metamask/src/index.ts
@@ -11,6 +11,7 @@ export const MetaMask: WalletAdapterType = ({ providersStore }) => ({
   walletName: name,
   walletId: id,
   type: injected.type,
+  rdns,
   icon: {
     light: WalletIcon,
     dark: WalletIconInverted,